+++ subterra-engine/src/main/java/io/toterra/subterra/engine/log/LogConfig.java
@@ -1,4 +1,4 @@
-package io.toterra.subterra.log;
+package io.toterra.subterra.engine.log;
 
 import io.toterra.subterra.engine.config.TdTable;
 import io.toterra.subterra.engine.config.TdValue;