+++ subterra-engine/src/main/java/io/toterra/subterra/engine/log/LogRing.java
@@ -1,4 +1,4 @@
-package io.toterra.subterra.log;
+package io.toterra.subterra.engine.log;
 
 import java.util.ArrayList;
 import java.util.List;